requirements.txt
tqdm
pandas
jupyter
notebook
streamlit
watchdog
plotly-express
numpy
